split components into separate jsx files
