+++ client/src/components/ratings-reviews/starbar.jsx
@@ -1,9 +1,31 @@
+import Star from '../../assets/Star.png';
+import quarterStar from '../../assets/quarterStar.png';
+import halfStar from '../../assets/halfStar.png';
+import threefourthsStar from '../../assets/3fourthsStar.png';
+import fullstar from '../../assets/fullstar.png';
 
-const Starbar = function ({rating}) {
+export const Starbar = function ({rating}) {
+
+  rating = ((rating / .25).toFixed(0)) * .25;
+  const ratingsArray = new Array(Math.floor(rating)).fill(fullstar);
+  rating = rating - Math.floor(rating);
+
+  if (rating === .75) {
+    ratingsArray.push(threefourthsStar)
+  } else if (rating === .5) {
+    ratingsArray.push(halfStar)
+  } else if (rating === .25) {
+    ratingsArray.push(quarterStar)
+  }
+  while (ratingsArray.length < 5) {
+    ratingsArray.push(Star)
+  }
 
   return (
     <div>
-
+      {ratingsArray.map((item, index)=> {
+        return <img key={index} src={item}></img>
+      })}
     </div>
   )
 }
